Update profile.xml
Hey CE0, 
Here is my version of Aela's profile. 
I tried to make her backstory as in line with the lore of Skyrim as possible and the rest, I just filled in with my imagination.
I don't know if you were looking for something this comprehensive when you said this would be a community space, but I just wanted to share what I did and to see what you thought about it.
I will be uploading profiles for other characters I've done as well.